requirements.txt
uvicorn
supabase
fastapi
pydantic
python-dotenv
python-jose
fastapi[all]
openai
watchgod
google-genai
pytest  
psycopg2-binary
pymemcache
redis 
geocoder
public-ip
pyjwt
gunicorn